requirements.txt
tensorflow==2.7.0
numpy
opencv-python
Pillow
deepface
keras
matplotlib
opencv-python
pyyaml